fix: automappers Dto to Model feat: returns erros in useCases refactor: result class properties refactor: all returns Result properties

+++ Parking.Management.Dashboard/Backend/src/domain/use-cases/vehicle/delete/delete-vehicle-usecase.ts
@@ -8,13 +8,21 @@ export class DeleteVehicleUseCase {
   constructor(private vehicleRepository: IVehicleRepository) {}
 
   async execute(id: string): Promise<Result<VehicleDto>> {
+    const vehicle = await this.vehicleRepository.findById(id);
+
+    if (!vehicle) {
+      return new Result<VehicleDto>({
+        message: "Veiculo não encontrado.",
+      });
+    }
+
     const result = await this.vehicleRepository.delete(id);
 
     const vehicleDto = mapper.map(result, Vehicle, VehicleDto);
 
-    return new Result<VehicleDto>(
-      vehicleDto,
-      "Universidade deletada com sucesso."
-    );
+    return new Result<VehicleDto>({
+      content: vehicleDto,
+      message: "Veiculo deletado com sucesso.",
+    });
   }
 }
